Added BLE api for f20
